+++ vllm/model_executor/layers/fused_moe/ampere_fp8_fused_moe.py
@@ -4,8 +4,6 @@ import functools
 import json
 import os
 from typing import Any, Dict, Optional, Tuple, Callable
-import pkg_resources
-import shutil
 
 import torch
 import triton
@@ -16,24 +14,6 @@ from vllm.logger import init_logger
 from vllm.utils import is_hip
 
 logger = init_logger(__name__)
-
-
-def replace_triton_cuda():
-
-    def get_package_path(package_name):
-        return pkg_resources.get_distribution(package_name).location
-
-    def get_package_version(package_name):
-        return pkg_resources.get_distribution(package_name).version
-    
-    assert get_package_version('triton') == "2.3.0" or get_package_version('triton') == "2.2.0"
-
-    cur_folder_cuda_py = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'triton_cuda.py')
-    target_folder_cuda_py = os.path.join(get_package_path('triton'), 'triton', 'language', 'extra', 'cuda.py')
-    shutil.copyfile(cur_folder_cuda_py, target_folder_cuda_py)
-
-
-replace_triton_cuda()
 
 
 @triton.jit
